requirements.txt
pip~=25.2
typing_extensions~=4.15.0
DateTime~=5.5
pytz~=2024.2
colorama~=0.4.6
sniffio~=1.3.1
kafka-python~=2.2.15
numpy~=2.3.1
setuptools~=75.6.0
regex~=2025.7.34
joblib~=1.5.1
pandas~=2.3.0
python-dateutil~=2.9.0.post0
six~=1.17.0
pydantic~=2.11.7
pydantic_core~=2.33.2
annotated-types~=0.7.0
starlette~=0.47.3
anyio~=4.10.0
nltk~=3.9.1
dotenv~=0.9.9
python-dotenv~=1.1.1
pymongo~=4.14.1
uvicorn~=0.35.0
fastapi~=0.116.1